+++ py_selenium_declarative/__init__.py
@@ -1,2 +1,2 @@
 __app_name__ = "py_selenium_declarative"
-__version__ = "0.1.0"
+__version__ = "{{version}}"