+++ py_selenium_declarative/model/suite.py
@@ -4,10 +4,10 @@ from py_selenium_declarative.model.operation import Operation
 
 
 class Suite:
-    name: str = ''
+    name: str = ""
     operations: List[Operation] = []
 
-    def __init__(self, operations, name=''):
+    def __init__(self, operations, name=""):
         self.name = name
         self.operations = []
         for operation in operations: