+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change log
 
+## [0.1.3] - 2023-08-12
+
+### Added
+- Basic Validator added.
+- Functionality added to verify text.
+- Tests and Coverage.
+
+### Modified
+- Moved driver related functionality to separate class.
+
 ## [0.1.2] - 2023-07-26
 
 ### Added